+++ component/common/relabel/relabel.go
@@ -186,7 +186,9 @@ func ComponentToPromRelabelConfigs(rcs []*Config) []*relabel.Config {
 }
 
 // Rules returns the relabel configs in use for a relabeling component.
-type Rules func() []*Config
+type Rules struct {
+	GetAll func() []*Config
+}
 
 // RiverCapsule marks the alias defined above as a "capsule type" so that it
 // cannot be invoked by River code.