+++ component/loki/relabel/relabel.go
@@ -58,8 +58,8 @@ func (a *Arguments) UnmarshalRiver(f func(interface{}) error) error {
 
 // Exports holds values which are exported by the loki.relabel component.
 type Exports struct {
-	Receiver loki.LogsReceiver  `river:"receiver,attr"`
-	Rules    flow_relabel.Rules `river:"rules,attr"`
+	Receiver loki.LogsReceiver   `river:"receiver,attr"`
+	Rules    *flow_relabel.Rules `river:"rules,attr"`
 }
 
 // Component implements the loki.relabel component.
@@ -98,7 +98,7 @@ func New(o component.Options, args Arguments) (*Component, error) {
 	// Create and immediately export the receiver which remains the same for
 	// the component's lifetime.
 	c.receiver = make(loki.LogsReceiver)
-	o.OnStateChange(Exports{Receiver: c.receiver, Rules: c.getRules})
+	o.OnStateChange(Exports{Receiver: c.receiver, Rules: getRules(c)})
 
 	// Call to Update() to set the relabelling rules once at the start.
 	if err := c.Update(args); err != nil {
@@ -233,9 +233,13 @@ func (c *Component) process(e loki.Entry) model.LabelSet {
 	return relabeled
 }
 
-func (c *Component) getRules() []*flow_relabel.Config {
-	c.mut.RLock()
-	defer c.mut.RUnlock()
+func getRules(c *Component) *flow_relabel.Rules {
+	return &flow_relabel.Rules{
+		GetAll: func() []*flow_relabel.Config {
+			c.mut.RLock()
+			defer c.mut.RUnlock()
 
-	return c.rcsFlow
+			return c.rcsFlow
+		},
+	}
 }